requirements.txt
click==8.1.7
python-docx==1.1.2
jinja2==3.1.4
ics==0.7.2
pytz==2024.1
ruamel.yaml==0.18.6
